Implemented ALL the tests.
Added "testing" and "nightly_testing" features to `zeroable_101`.

Conditionally enable nightly features for tests,
when tests are run with the "nightly_testing" feature.

Added "regex" as a dev-dependency of `zeroable_101_derive`

Added a `#[zero(_test_code="")]` attribute to inject code in tests.

Fixed handling of zero field unions to error.